HMP-437: restores custom fonts, refactors overridden components
