+++ ucb_extras/bampfa/app/assets/stylesheets/extras.scss
@@ -71,31 +71,31 @@
 
 @font-face {
   font-family: RobotoBlack;
-  src: url(/fonts/Roboto/Roboto-Black.ttf);
+  src: url(Roboto/Roboto-Black.ttf);
   font-weight: bold;
 }
 
 @font-face {
   font-family: Roboto;
-  src: url(/fonts/Roboto/Roboto-Regular.ttf);
+  src: url(Roboto/Roboto-Regular.ttf);
   font-weight: normal;
 }
 
 @font-face {
   font-family: RobotoMedium;
-  src: url(/fonts/Roboto/Roboto-Medium.ttf);
+  src: url(Roboto/Roboto-Medium.ttf);
   font-weight: normal;
 }
 
 @font-face {
   font-family: Roboto;
-  src: url(/fonts/Roboto/Roboto-Bold.ttf);
+  src: url(Roboto/Roboto-Bold.ttf);
   font-weight: bold;
 }
 
 @font-face {
   font-family: RobotoLight;
-  src: url(/fonts/Roboto/Roboto-Light.ttf);
+  src: url(Roboto/Roboto-Light.ttf);
   font-weight: normal;
 }
