+++ ucb_extras/cinefiles/app/assets/stylesheets/extras.scss
@@ -160,31 +160,31 @@ article .documentHeader {
 
 @font-face {
   font-family: MontserratBlack;
-  src: url(/fonts/Montserrat/Montserrat-Black.ttf);
+  src: url(Montserrat/Montserrat-Black.ttf);
   font-weight: bold;
 }
 
 @font-face {
   font-family: Montserrat;
-  src: url(/fonts/Montserrat/Montserrat-Regular.ttf);
+  src: url(Montserrat/Montserrat-Regular.ttf);
   font-weight: normal;
 }
 
 @font-face {
   font-family: MontserratMedium;
-  src: url(/fonts/Montserrat/Montserrat-Medium.ttf);
+  src: url(Montserrat/Montserrat-Medium.ttf);
   font-weight: normal;
 }
 
 @font-face {
   font-family: Montserrat;
-  src: url(/fonts/Montserrat/Montserrat-Bold.ttf);
+  src: url(Montserrat/Montserrat-Bold.ttf);
   font-weight: bold;
 }
 
 @font-face {
   font-family: MontserratLight;
-  src: url(/fonts/Montserrat/Montserrat-Light.ttf);
+  src: url(Montserrat/Montserrat-Light.ttf);
   font-weight: normal;
 }
