--- ucb_extras/common/app/components/blacklight/constraint_component.html.erb
@@ -1,6 +1,0 @@
-<%= render(@layout.new(
-      index: @index,
-      classes: (Array(@classes) + ["filter-#{@facet_item_presenter.key.parameterize}"]).join(' '),
-      label: @facet_item_presenter.field_label,
-      value: @facet_item_presenter.constraint_label,
-      remove_path: @facet_item_presenter.remove_href)) %>
