--- ucb_extras/common/app/components/blacklight/constraint_layout_component.html.erb
@@ -1,24 +1,0 @@
-<span class="btn-group applied-filter constraint <%= @classes %>">
-  <span class="constraint-value btn btn-outline-secondary">
-    <% if @label.present? %>
-      <span class="filter-name"><%= @label %></span>
-    <% end %>
-    <% if @value.present? %>
-      <%= content_tag :span, @value, class: 'filter-value' %>
-    <% end %>
-  </span>
-  <% if @remove_path.present? %>
-    <%= link_to(
-        helpers.with_screen_reader_alert(
-          @remove_path,
-          "Removed #{@label}: \"#{@value}\" from search constraints",
-          @next_focus_target
-        ),
-        class: 'btn btn-outline-secondary remove',
-        id: "remove-constraint-#{@index}",
-    ) do %>
-      <%= render Blacklight::Icons::RemoveComponent.new %>
-      <span class="sr-only visually-hidden"><%= remove_aria_label %></span>
-    <% end %>
-  <% end %>
-</span>
