+++ ucb_extras/common/app/components/response/facet_group_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Response
+  # Render a group of facet fields
+  class FacetGroupComponent < Blacklight::Response::FacetGroupComponent
+
+    def should_collapse_facets?
+      helpers.search_state.filters.blank?
+    end
+  end
+end
