+++ ucb_extras/common/app/views/shared/_flash_msg.html.erb
@@ -1,5 +1,5 @@
 <div class="flash_messages" role="status">
   <% [:success, :notice, :error, :alert, :sr_alert].each do |type| %>
-    <%= render(Blacklight::System::FlashMessageComponent.with_collection(Array.wrap(flash[type]), type: type)) if flash[type] %>
+    <%= render(System::FlashMessageComponent.with_collection(Array.wrap(flash[type]), type: type)) if flash[type] %>
   <% end %>
 </div>
