+++ ucb_extras/common/config/initializers/assets.rb
@@ -5,10 +5,11 @@ Rails.application.config.assets.version = "1.0"
 
 # Add additional assets to the asset load path.
 Rails.application.config.assets.paths <<  Rails.root.join('app', 'assets', 'stylesheets').to_s
+Rails.application.config.assets.paths <<  Rails.root.join('app', 'assets', 'fonts').to_s
+Rails.application.config.assets.paths << Rails.root.join('node_modules/openseadragon/build/openseadragon/images')
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
 # Rails.application.config.assets.precompile += %w[ admin.js admin.css ]
 
-Rails.application.config.assets.paths << Rails.root.join('node_modules/openseadragon/build/openseadragon/images')
